Fix smooth fork initial file selection on live
