+++ packages/app/src/app/overmind/namespaces/editor/actions.ts
@@ -10,6 +10,7 @@ import {
 import { Action, AsyncAction } from 'app/overmind';
 import { withLoadApp, withOwnedSandbox } from 'app/overmind/factories';
 import { sortObjectByKeys } from 'app/overmind/utils/common';
+import getItems from 'app/overmind/utils/items';
 import {
   addDevToolsTab as addDevToolsTabUtil,
   closeDevToolsTab as closeDevToolsTabUtil,
@@ -101,6 +102,11 @@ export const sandboxChanged: AsyncAction<{ id: string }> = withLoadApp<{
     await effects.vscode.closeAllTabs();
 
     actions.internal.setCurrentSandbox(sandbox);
+
+    const items = getItems(state);
+    const defaultItem = items.find(i => i.defaultOpen) || items[0];
+
+    state.workspace.openedWorkspaceItem = defaultItem.id;
   } catch (error) {
     state.editor.notFound = true;
     state.editor.error = error.message;
